requirements.txt
Pillow
numpy
torch
filetype
gradio
gradio-rangeslider
onnx
onnxruntime
onnxruntime-gpu
opencv-python
psutil
tqdm
scipy
